Don't Like data is also stored at the Show level now
Don't Like behaves similar to Watched now in that if you mark one Airing
of something as Don't Like; then that applies to all other Airings of the
same Show. This shouldn't affect Intelligent Recording at all; but it will
prevent the Scheduler from putting this in the schedule; even if they are
Favorites. As part of this; the Scheduler will no longer schedule anything
that is Don't Like (it sort of did this before, but not very effectively).
The data is encoded into the Show.lastWatched field by using negative values
so this did not introduce any kind of database incompatability. The propagation to
Show will also only apply to newly set Don't Likes.

There was also some other random code cleanup along the way.