Change keyword Favorite behavior to be like it was in SageTV 7
In V9, Lucene was being used for keyword Favorite matching which essentially disabled
the use of regex for keywords. For compatibility, by default I'm changing it back to
how it did things in V7. This can be changed to use the new method by setting the 
property "use_legacy_keyword_favorites=false" on the server. This only impacts
performance if keyword favorites actually exist; and they will be faster with the
new method although it loses the regex feature.