Updated system message handling for HALT Detected messages
There's a new property "msg/halt_count_for_error" (default 2) which
allows controlling how many consecutive HALTs occur before it goes
from the warning to the error state for the message. I've also extended
the "msg/alert_enabled/*" property so that it can also do
"msg/alert_enabled/[type]-[priority]" to enable raising the global alert
level for only a certain priority of message. This should have no change
to the currently configured behavior for anybody.

What this enables is for showing a global alert in the UI if there's
more than X consecutive halts on an encoder. Personally, I have this
problem with my HDPVRs when they get into an error state and this will
allow me to easily notice it when it occurs rather than days later when
I try to play an affected recording.