Added new system message for when a recording's bitrate is too low
This is disabled by default. But a property can be set for each capture
device using the format
mmc/encoders/XXXX/minimum_bitrate_error_threshold_bps which is in bits
per second. If a recording on that capture device goes below that
bitrate, then it will fire a system error message. The value default to
zero, which means there is no low bitrate detection for that encoder.

The reason I added this is because it's a reliable way for me to detect
when my cable boxes are off and the HDPVR is just recording a black
screen from them.