Fixed issues with loading channel logos on the UI thread
There were various places where the existence of a Channel
logo were being determined by querying for the image itself. Normally,
this is fine if it is being done leading up to an Image or Text widget
as those will be properly backgrounded, but some were leading up to
Panel widgets. I changed those to check the URL rather than the image
which will not result in any fetching of remote data in-line.