Worked around issue where sometimes TV recordings were detected as audio only
The actual bug here is in the native format detector. However, after analyzing that, it ended up
being a corner case where a certain stream type in the TS map (DC-II) was being considered
video if it was a TV recording and audio otherwise because both things existed in the wild. Changing
that behavior would likely have unintended side effects.

Normally, it worked around this problem because it would get a hint that something was a TV
recording and then use that to prefer video when it saw the DC-II marker rather than audio. However,
that hint was not being set when you were using a network encoder because then the format
detection was being done through a higher level mechanism.

The workaround, which defaults to being on via the property "double_check_audioonly_tv_formats" will
redo the format detection with FFMPEG when it sees that only an audio stream was detected in an
MPEG file. If FFMPEG also does not detect a video stream, then it will use the originally detected internal
format; but if FFMPEG does detect a video stream, then it will return FFMPEG's format instead.

I don't see any potential negative effects from making this change, but please let me know if
anything comes up.