Added 'static' keyword because it was crashing without it when running
Also removing the 'inline' keyword fixed it as well, but I didn't do that.
This clearly looks like a compiler error even though I saw it with two
different versions of GCC. Looking at the disassembly showed that it was
still trying to call an IsNTSCVideo function even though it was not
defined. Others indicated this is not a problem...but it won't hurt
anybody else and I just built this on a new copy of Ubuntu from last week,
so hopefully it'll help somebody out. :)