’Test99: Namé’ is becoming ’Test99: NamÃ©’ on Windows.
The test isn't working on Windows. ’Test99: Namé’ is being turned into ’Test99: NamÃ©’ by the compiler. I looked up the UTF-8  code for this the character é and it's 0xC3A9 which explains how one character somehow became two (even though that should fit within a character; maybe a bug in Java too). Oddly enough in UTF-16, it's 0x00E9, so I change é to \u00E9 and now the test is passing.