Fixed bug where we could not do live preview always
This affected CaptureDevices that were not yet configured.
There was new code added that had an NPE in this case,
and that's resolved now. This manifested itself as live
preview never working in the Setup Wizard after first
selecting a capture device if it had not yet been configured.
Also caused issues with detecting the broadcast standard.