Fixed rewind issue
This was a problem where if you were rewinding to the beginning of a recording
it would end up skipping to the end right before you got to the beginning. The bug
was that it was then using the seeked to PTS as a parsed one, which was less
than the firstPTS in the file so it flagged it with ptsRollover and ended up
wrapping that value around which ended up being a very large value at the end
of the file. By clamping the seek input, we avoid using that bad PTS value.