Byte based seeking support for MPEG files
This helps to resolve a long standing issue where files that have bad PTS timelines
cause the playback time to jump wildly and infuriate you when you try
to skip around them. :)

What it does is detect when the parser's calculated duration for a file that is
being played back (extender only, no transcoding) is different by a decent amount (25%)
from that of what SageTV thinks the recorded duration should be. Then it uses the read
file position to estimate the current playback time and then for seeking, it guesses
the proper file position as if the file were constant bitrate. This works WAY better
than being stuck without the ability to seek or even know what time you are at in the file.

This feature is off by default, and it can be enabled by setting this in your Sage.properties file:
disable_byte_based_seek_check=false

This does NOT work on files that are currently recording.
This does NOT allow you to fast forward or rewind properly in files like this, only seek.
You CAN force this to work for currently recording files by setting this property in the
properties file for the extender you want it to work on:
force_byte_based_seeking=true
(this is so if you're recording something you really want to watch right now that has
this issue, you can turn off the extender, edit the properties file and then it'll make it
happen...don't forget to reset that property later).
This is somewhat inaccurate, but that's what you get with byte-based time estimation.