Fixed Wasted cleanup and 4K imagery extender crash
This implemented a change that was documented but never done.
For Wasted objects that are automatically created (i.e. the user did not select Don't Like)
these are now deleted after they are over a year old. This will reduce database size.

When we were using diffused images on the extender (like the reflection you see
in the guide and detailed info page), we would forcibly load the image at
full size to workaround specifics in the background image loading algorithms.
We still forcibly load an image, but now it ensures it does not exceed the maximum
image size we use (1024).